Use readme as a starting point for users
This pull request is to make the cabs/README.md the starting point for new CABS users. The thought is we make a new Launchpad team for the perspective CPP partner, we add that team to CABS, and then direct them to this readme for directions on usage.  The readme should evolve with the project and thus should live in the top level project.

Note this pull request should be validated by someone other than me.

-thanks,
Antonio